Formatting with Perltidy
git-svn-id: svn://svn.code.sf.net/p/ledger-smb/code/branches/1.2@1068 4979c152-3d1c-0410-bac9-87ea11338e46

+++ sql/Pg-database.sql
@@ -102,7 +102,7 @@ CREATE TABLE invoice (
   parts_id int,
   description text,
   qty NUMERIC,
-  allocated integer,
+  allocated NUMERIC,
   sellprice NUMERIC,
   fxsellprice NUMERIC,
   discount float4, -- jd: check into this
